pyproject.toml
[tool.poetry]
name = "test-task-for-standart"
version = "0.1.0"
description = ""
authors = ["odnashestaia <odna_shestaia@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
flask-wtf = "^1.2.2"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
celery = {extras = ["beat"], version = "^5.4.0"}
flask-restful = "^0.3.10"
flask-admin = "^1.6.1"
flask-migrate = "^4.0.7"
requests = "^2.32.3"
django-environ = "^0.11.2"
flasgger = "^0.9.7.1"
flower = "^2.0.1"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
kombu = "^5.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
flask = "app:app.run"
celery_worker = "celery -A app.celery worker --loglevel=info"
celery_beat = "celery -A app.celery beat --loglevel=info"